pyproject.toml
[tool.poetry]
name = "dmlgames"
version = "0.1.0"
description = ""
repository = "https://github.com/Yobmod/dmlgames"
authors = ["yobmod <yobmod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pygame = "2.0.0.dev8"
kivy = "*"
kivy-examples = "^1.11.1"
kivy-garden = "^0.1.4"
"kivy_deps.sdl2" = "0.1.*"
"kivy_deps.glew" = "0.1.*"
"kivy_deps.angle" = "0.1.*"
"kivy_deps.gstreamer" = "0.1.*"
typing-extensions = "*"
# # boltons = "*"
# colorama = "*"
# termcolor = "*"
# chempy = "*"
# pyyaml = "*"
# returns = "*"
# typer = {extras = ["all"], version = "*"}
# dacite = "*"
# dill = "*"
# toml = "*"
# pillow = "*"
# imutils = "*"
# mutagen = "*"
# boltons = "*"
# pyyaml = "*"
# sqlalchemy-stubs = "*"
# rapidtables = "*"
# ordered-set = "*"
# ordered-set-stubs = "*"
# toolz = "*"
# cytoolz = "*"


[tool.poetry.dev-dependencies]
pytest = "*"
# cython = "*"
mypy = "*"
pytest-cov = "*"  # brings coverage.py
pytest-mypy = "*"
pytest-sugar = "*"
pytest-xdist = "*"
pylama = "*"
autopep8 = "*"
flake8 = "*"
flake8-comprehensions = "*"
flake8-bugbear = "*"
pdbpp = "*"
pyinstaller = "*"
# safety = "*"
# black = "^19.10b0"
gitpython = "*"
python-dotenv = "*"

[settings]
pythonPaths = ["C:\\Users\\Customer\\AppData\\Local\\pypoetry\\Cache\\virtualenvs\\dmlapp-SA54dGKK-py3.7",
                "C:\\Users\\Yobmod\\AppData\\Local\\pypoetry\\Cache\\virtualenvs\\dmlgames-wH0R5DCa-py3.7"]


[build-system]
requires = ["poetry>=1.05"]
build-backend = "poetry.masonry.api"
